magic numbers are now encoded as Long for very fast explorations

+++ oscar-algo/src/main/java/oscar/algo/reversible/Reversible.java
@@ -21,7 +21,7 @@ package oscar.algo.reversible;
  */
 public abstract class Reversible {
 	
-	private int lastMagic;
+	private long lastMagic;
 	public final ReversibleContext node;
 	
 	public Reversible(ReversibleContext node) {
